Docs (article): Update MCP integration documentation and sample code structure
-Simplify the MCP document content and highlight the integration of HTTP tools and NPX servers
-Update configuration examples and code snippets to optimize document readability
-Adjust the citation numbers of MCP related articles and reorganize the learning path
-Remove the example code of dual Agent concatenation in Article 08Mcp
-Correct the class name and file reference path in the flight comparison example
-Delete redundant MCP Function Rolling and AgentExecutor article content
